Fixed bug in selecting candidates from bins
With certain unlucky bin sizes, the selection pool for the last part of
the trajectory could become very small. If the last bin in the database
is too small, it gets merged with the previous bin. However, a bug
caused all, except the first day in the last bin, to be merged. This
created a very small bin, which (if the last segment in the trajectory
coincidentally started on this day) resulted in a very small search pool
on the last segment